Fix content views not updating
Content views weren't getting marked dirty inherently on notification
updates from the app, so the inflater wouldn't bother binding the
content view again, even if the layout actually changed. This fixes
that by providing the ability to mark all the active content views as
dirty and needing to be rebound and calling it on add/update.

Bug: 149002542
Test: repro from bug
Test: atest RowContentBindStageTest
Change-Id: I59d7f86e43763c3c03329a60a052e6c51cdc845a