Revert "Turn on QS media player by default"
This reverts commit 80504fec1b8f9c11d5fe6bcc4135f3bb1426b2ff.

Reason for revert: b/150437753. IM'ed with Beth, going to revert but not merge this CL.  Beth is working on a fix with an eta of a day or 2. Ill CP this into a RP1A.200303.002+ builds.

Change-Id: Id42659d33bbd5ccc188d977f6af320cc8b3b6478
(cherry picked from commit f1563675e81b0b079b90fa3d7c9628c16aab9359)