Limit framework module api packages
Use the same whitelist of packages that the non-updatable
part of the platform uses for now. This avoids exposing some
internal classes in a follow-up CL.

Move telephony off the module-wide defaults for now, as the
telephony stubs are causing problems.

Bug: 149293194
Test: m checkapi
Change-Id: Iedc8c309cef8290ac14e1ad32b5b66656145f557
Merged-In: Iedc8c309cef8290ac14e1ad32b5b66656145f557